@@ -1,12 +1,13 @@
 import 'package:dart_frog/dart_frog.dart';
 import 'package:postgres/postgres.dart';
 
+import '../../../core/db_helpers.dart';
 import '../models/db_transaction.dart';
 
 Middleware transactionRepositoryProvider() {
   return provider<Future<TransactionRepository>>(
     (context) async => TransactionRepository(
-      await context.read<Future<PostgreSQLConnection>>(),
+      await context.read<Future<Connection>>(),
     ),
   );
 }
@@ -14,43 +15,43 @@ Middleware transactionRepositoryProvider() {
 class TransactionRepository {
   TransactionRepository(this.connection);
 
-  final PostgreSQLConnection connection;
+  final Connection connection;
 
   /// Get a paginated list of transactions.
   Future<List<DbTransaction>> getTransactions(String userId, int offset, [int limit = 20]) async {
-    final result = await connection.mappedResultsQuery(
+    final result = await connection.executeNamed(
       'SELECT * FROM transactions WHERE user_id = @userId ORDER BY date DESC LIMIT @limit OFFSET @offset',
-      substitutionValues: {
+      parameters: {
         'userId': userId,
         'limit': limit,
         'offset': offset,
       },
     );
 
-    return result.map((e) => DbTransaction.fromDatabase(e['transactions']!)).toList();
+    return result.map((e) => DbTransaction.fromDatabase(e.toColumnMap())).toList();
   }
 
   /// Get a single transaction.
   Future<DbTransaction> getTransaction(String userId, String transactionId) async {
-    final result = await connection.mappedResultsQuery(
+    final result = await connection.executeNamed(
       'SELECT * FROM transactions WHERE user_id = @userId AND id = @transactionId',
-      substitutionValues: {
+      parameters: {
         'userId': userId,
         'transactionId': transactionId,
       },
     );
 
-    return DbTransaction.fromDatabase(result.first['transactions']!);
+    return DbTransaction.fromDatabase(result.first.toColumnMap());
   }
 
   /// Create a new transaction and return the result.
   Future<DbTransaction> createTransaction(DbTransaction transaction) async {
-    final result = await connection.mappedResultsQuery(
+    final result = await connection.executeNamed(
       'INSERT INTO transactions'
       '(user_id, date, text, amount, account_id, category_id, fixed_cost, description) VALUES'
       '(@userId, @date, @text, @amount, @accountId, @categoryId, @fixedCost, @description)'
       'RETURNING *',
-      substitutionValues: {
+      parameters: {
         'userId': transaction.userId,
         'date': transaction.date,
         'text': transaction.text,
@@ -62,11 +63,11 @@ class TransactionRepository {
       },
     );
 
-    return DbTransaction.fromDatabase(result.first['transactions']!);
+    return DbTransaction.fromDatabase(result.first.toColumnMap());
   }
 
   Future<DbTransaction> updateTransaction(DbTransaction transaction) async {
-    final result = await connection.mappedResultsQuery(
+    final result = await connection.executeNamed(
       'UPDATE transactions SET '
       'date = @date, '
       'text = @text, '
@@ -77,7 +78,7 @@ class TransactionRepository {
       'description = @description '
       'WHERE id = @id AND user_id = @userId '
       'RETURNING *',
-      substitutionValues: {
+      parameters: {
         'id': transaction.id,
         'userId': transaction.userId,
         'date': transaction.date,
@@ -90,13 +91,13 @@ class TransactionRepository {
       },
     );
 
-    return DbTransaction.fromDatabase(result.first['transactions']!);
+    return DbTransaction.fromDatabase(result.first.toColumnMap());
   }
 
   Future<void> deleteTransaction(String userId, String transactionId) async {
-    await connection.execute(
+    await connection.executeNamed(
       'DELETE FROM transactions WHERE user_id = @userId AND id = @transactionId',
-      substitutionValues: {
+      parameters: {
         'userId': userId,
         'transactionId': transactionId,
       },
